+++ backend/src/baserow/core/management/backup/exceptions.py
@@ -1,0 +1,4 @@
+class InvalidBaserowBackupArchive(Exception):
+    """
+    Raised when a restore is attempted with an invalid Baserow backup archive.
+    """